01-Agentic-AI-Workflow/pyproject.toml
[project]
name = "01-agentic-ai-workflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.47.0",
    "huggingface-hub<1.0",
    "openai>=2.30.0",
    "pypdf>=6.9.2",
    "requests>=2.32.5",
    "ruff>=0.15.9",
]
